Integrate A/B switching
Add calls to tools abctl/qbootctl and call when appropriate. If this is
running as osbuild just set up the correct symlink. Otherwise read
cmdline to determine what the new slot is. The location of the new root
directory is passed via ostree. All this occurs during shutdown at the
"OSTree Finalize Staged Deployment" part of the deploy, with the
exception of when this is called during osbuild.